More updates to make work
Still fighting with the disk managemention.  This version appears to work and the logging is helpful for determining what is going on with the partition blocks.

Note teat resize2fs seems to reboot the computer one, so expect it to go through that a second time (hence all the if else statements) and continue the disk managment after a reboot.